# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RelayStats {
  providers: number;
  activeKeys: number;
  requestsRouted: number;
  uptimePercent: number;
  endpoint: string;
}

export interface RelayProvider {
  id: string;
  name: string;
  shortName: string;
  kind: string;
  model: string;
  endpoint: string;
  keyCount: number;
  status: string;
  accent: string;
}

export interface RelayKeyInput {
  providerId: string;
  /** @minLength 8 */
  apiKey: string;
  /** @maxLength 48 */
  label?: string;
  consent: boolean;
}

export interface RelayKey {
  id: string;
  providerId: string;
  label: string;
  maskedKey: string;
  status: string;
  createdAt: string;
}

export interface RelayMessage {
  role: string;
  content: string;
}

export interface RelayChatInput {
  /** @minItems 1 */
  messages: RelayMessage[];
  model?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
}

export type RelayChatResponseChoicesItemMessage = {
  role: string;
  content: string;
};

export type RelayChatResponseChoicesItem = {
  index: number;
  message: RelayChatResponseChoicesItemMessage;
  finish_reason: string;
};

export interface RelayChatResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: RelayChatResponseChoicesItem[];
  provider: string;
  content: string;
  routedAt: string;
  fallbackCount: number;
}

export interface ErrorResponse {
  error: string;
}