# lib/api-zod/src/generated/types/relayChatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelayMessage } from './relayMessage';

export interface RelayChatInput {
  /** @minItems 1 */
  messages: RelayMessage[];
  model?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
}