# lib/api-zod/src/generated/types/relayChatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelayChatResponseChoicesItem } from './relayChatResponseChoicesItem';

export interface RelayChatResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: RelayChatResponseChoicesItem[];
  provider: string;
  content: string;
  routedAt: string;
  fallbackCount: number;
}