# lib/api-zod/src/generated/types/relayKey.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayKey {
  id: string;
  providerId: string;
  label: string;
  maskedKey: string;
  status: string;
  createdAt: string;
}