# lib/api-zod/src/generated/types/relayKeyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayKeyInput {
  providerId: string;
  /** @minLength 8 */
  apiKey: string;
  /** @maxLength 48 */
  label?: string;
  consent: boolean;
}