# lib/api-zod/src/generated/types/relayProvider.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayProvider {
  id: string;
  name: string;
  shortName: string;
  kind: string;
  model: string;
  endpoint: string;
  keyCount: number;
  status: string;
  accent: string;
}