# lib/api-zod/src/generated/types/relayStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayStats {
  providers: number;
  activeKeys: number;
  requestsRouted: number;
  uptimePercent: number;
  endpoint: string;
}